Organise ingestion module
