+++ ingestion/src/metadata/ingestion/source/database/bigquery.py
@@ -43,9 +43,9 @@ from metadata.generated.schema.metadataIngestion.workflow import (
 from metadata.generated.schema.type.tagLabel import TagLabel
 from metadata.ingestion.api.source import InvalidSourceException
 from metadata.ingestion.models.ometa_tag_category import OMetaTagAndCategory
+from metadata.ingestion.source.database.column_type_parser import create_sqlalchemy_type
 from metadata.ingestion.source.database.common_db_source import CommonDbSourceService
 from metadata.utils import fqn
-from metadata.utils.column_type_parser import create_sqlalchemy_type
 from metadata.utils.logger import ingestion_logger
 
 logger = ingestion_logger()
