+++ ingestion/src/metadata/ingestion/source/database/dbt_source.py
@@ -24,8 +24,8 @@ from metadata.generated.schema.entity.data.table import (
 from metadata.generated.schema.type.entityLineage import EntitiesEdge
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
+from metadata.ingestion.source.database.column_type_parser import ColumnTypeParser
 from metadata.utils import fqn
-from metadata.utils.column_type_parser import ColumnTypeParser
 from metadata.utils.logger import ingestion_logger
 
 logger = ingestion_logger()
