+++ ingestion/src/metadata/ingestion/source/pipeline/airbyte.py
@@ -16,6 +16,7 @@ from typing import Iterable, Optional
 
 from pydantic import BaseModel
 
+from metadata.clients.airbyte_client import AirbyteClient
 from metadata.generated.schema.api.data.createPipeline import CreatePipelineRequest
 from metadata.generated.schema.api.lineage.addLineage import AddLineageRequest
 from metadata.generated.schema.entity.data.pipeline import (
@@ -41,7 +42,6 @@ from metadata.ingestion.api.source import InvalidSourceException
 from metadata.ingestion.models.pipeline_status import OMetaPipelineStatus
 from metadata.ingestion.source.pipeline.pipeline_service import PipelineServiceSource
 from metadata.utils import fqn
-from metadata.utils.airbyte_client import AirbyteClient
 from metadata.utils.logger import ingestion_logger
 
 logger = ingestion_logger()
