+++ ingestion/tests/unit/test_column_type_parser.py
@@ -1,7 +1,7 @@
 import os
 from unittest import TestCase
 
-from metadata.utils.column_type_parser import ColumnTypeParser
+from metadata.ingestion.source.database.column_type_parser import ColumnTypeParser
 
 COLUMN_TYPE_PARSE = [
     "array<string>",
